requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
requests==2.7.0
SQLAlchemy==1.0.4
Werkzeug==0.10.4
